adding things slowly
